+++ js/ui.modal.js
@@ -14,13 +14,15 @@ var supportTransition = UI.support.transition;
 var Modal = function(element, options) {
   this.options = $.extend({}, Modal.DEFAULTS, options || {});
   this.$element = $(element);
+  this.$dialog =   this.$element.find('.am-modal-dialog');
 
   if (!this.$element.attr('id')) {
     this.$element.attr('id', UI.utils.generateGUID('am-modal'));
   }
 
   this.isPopup = this.$element.hasClass('am-popup');
-  this.active = this.transitioning = null;
+  this.isActions = this.$element.hasClass('am-modal-actions');
+  this.active = this.transitioning = this.relatedTarget = null;
 
   this.events();
 };
@@ -34,23 +36,29 @@ Modal.DEFAULTS = {
     modal: '.am-modal',
     active: '.am-modal-active'
   },
+  closeViaDimmer: true,
   cancelable: true,
   onConfirm: function() {
   },
   onCancel: function() {
   },
+  height: undefined,
+  width: undefined,
   duration: 300, // must equal the CSS transition duration
   transitionEnd: supportTransition && supportTransition.end + '.modal.amui'
 };
 
-Modal.prototype.toggle = function(relatedElement) {
-  return this.active ? this.close() : this.open(relatedElement);
+Modal.prototype.toggle = function(relatedTarget) {
+  return this.active ? this.close() : this.open(relatedTarget);
 };
 
-Modal.prototype.open = function(relatedElement) {
+Modal.prototype.open = function(relatedTarget) {
   var $element = this.$element;
   var options = this.options;
   var isPopup = this.isPopup;
+  var width = options.width;
+  var height = options.height;
+  var style = {};
 
   if (this.active) {
     return;
@@ -59,6 +67,9 @@ Modal.prototype.open = function(relatedElement) {
   if (!this.$element.length) {
     return;
   }
+
+  // callback hook
+  relatedTarget && (this.relatedTarget = relatedTarget);
 
   // 判断如果还在动画，就先触发之前的closed事件
   if (this.transitioning) {
@@ -71,15 +82,34 @@ Modal.prototype.open = function(relatedElement) {
 
   this.active = true;
 
-  $element.trigger($.Event('open.modal.amui', {relatedTarget: relatedElement}));
+  $element.trigger($.Event('open.modal.amui', {relatedTarget: relatedTarget}));
 
   dimmer.open($element);
 
   $element.show().redraw();
 
-  !isPopup && $element.css({
-    marginTop: -parseInt($element.height() / 2, 10) + 'px'
-  });
+  // apply Modal width/height if set
+  if (!isPopup && !this.isActions) {
+    if (width) {
+      width = parseInt(width, 10);
+      style.width =  width + 'px';
+      style.marginLeft =  -parseInt(width / 2) + 'px';
+    }
+
+    if (height) {
+      height = parseInt(height, 10);
+      // style.height = height + 'px';
+      style.marginTop = -parseInt(height / 2) + 'px';
+
+      // the background color is styled to $dialog
+      // so the height should set to $dialog
+      this.$dialog.css({height: height + 'px'});
+    } else {
+      style.marginTop = -parseInt($element.height() / 2, 10) + 'px';
+    }
+
+    $element.css(style);
+  }
 
   $element.
     removeClass(options.className.out).
@@ -89,7 +119,7 @@ Modal.prototype.open = function(relatedElement) {
 
   var complete = function() {
     $element.trigger($.Event('opened.modal.amui',
-      {relatedTarget: relatedElement}));
+      {relatedTarget: relatedTarget}));
     this.transitioning = 0;
   };
 
@@ -102,7 +132,7 @@ Modal.prototype.open = function(relatedElement) {
     emulateTransitionEnd(options.duration);
 };
 
-Modal.prototype.close = function(relatedElement) {
+Modal.prototype.close = function(relatedTarget) {
   if (!this.active) {
     return;
   }
@@ -120,7 +150,7 @@ Modal.prototype.close = function(relatedElement) {
   }
 
   this.$element.trigger($.Event('close.modal.amui',
-    {relatedTarget: relatedElement}));
+    {relatedTarget: relatedTarget}));
 
   this.transitioning = 1;
 
@@ -158,46 +188,60 @@ Modal.prototype.events = function() {
           this.close();
         }
       }, that));
+  }
 
-    dimmer.$element.on('click', function(e) {
+  // Close Modal when dimmer clicked
+  if (this.options.closeViaDimmer) {
+    dimmer.$element.on('click.dimmer.modal.amui', function(e) {
       that.close();
     });
   }
 
   // Close button
-  $element.find('[data-am-modal-close]').on('click.modal.amui', function(e) {
+  $element.find('[data-am-modal-close], .am-modal-btn').
+    on('click.close.modal.amui', function(e) {
     e.preventDefault();
     that.close();
   });
 
-  $element.find('.am-modal-btn').on('click.modal.amui', function(e) {
-    that.close();
-  });
+  // get inputs data
+  function getData() {
+    var data = [];
 
-  $element.find('[data-am-modal-confirm]').on('click.modal.amui', function() {
-    that.options.onConfirm($ipt.val());
-  });
+    $ipt.each(function() {
+      data.push($(this).val());
+    });
 
-  $element.find('[data-am-modal-cancel]').on('click.modal.amui', function() {
-    that.options.onCancel($ipt.val());
+    return (data.length === 0) ? undefined :
+      ((data.length === 1) ? data[0] : data);
+  }
+
+  $element.find('[data-am-modal-confirm]').on('click.confirm.modal.amui',
+    function() {
+      that.options.onConfirm.call(that, {data: getData()});
+    });
+
+  $element.find('[data-am-modal-cancel]').on('click.cancel.modal.amui',
+    function() {
+    that.options.onCancel.call(that, {data: getData()});
   });
 };
 
-function Plugin(option, relatedElement) {
+function Plugin(option, relatedTarget) {
   return this.each(function() {
     var $this = $(this);
-    var data = $this.data('am.modal');
+    var data = $this.data('amui.modal');
     var options = $.extend({},
       Modal.DEFAULTS, typeof option == 'object' && option);
 
     if (!data) {
-      $this.data('am.modal', (data = new Modal(this, options)));
+      $this.data('amui.modal', (data = new Modal(this, options)));
     }
 
     if (typeof option == 'string') {
-      data[option](relatedElement);
+      data[option](relatedTarget);
     } else {
-      data.toggle(option && option.relatedElement || undefined);
+      data.toggle(option && option.relatedTarget || undefined);
     }
   });
 }
@@ -210,7 +254,7 @@ $doc.on('click.modal.amui.data-api', '[data-am-modal]', function() {
   var options = UI.utils.parseOptions($this.attr('data-am-modal'));
   var $target = $(options.target ||
   (this.href && this.href.replace(/.*(?=#[^\s]+$)/, '')));
-  var option = $target.data('am.modal') ? 'toggle' : options;
+  var option = $target.data('amui.modal') ? 'toggle' : options;
 
   Plugin.call($target, option, this);
 });