+++ js/ui.popover.js
@@ -19,6 +19,7 @@ var Popover = function(element, options) {
 };
 
 Popover.DEFAULTS = {
+  theme: undefined,
   trigger: 'click',
   content: '',
   open: false,
@@ -182,12 +183,20 @@ Popover.prototype.close = function() {
 
 Popover.prototype.getPopover = function() {
   var uid = UI.utils.generateGUID('am-popover');
-  return $(this.options.tpl).attr('id', uid);
+  var theme = [];
+
+  if (this.options.theme) {
+    $.each(this.options.theme.split(','), function(i, item) {
+      theme.push('am-popover-' + $.trim(item));
+    });
+  }
+  return $(this.options.tpl).attr('id', uid).addClass(theme.join(' '));
 };
 
-Popover.prototype.setContent = function() {
+Popover.prototype.setContent = function(content) {
+  content = content || this.options.content;
   this.$popover && this.$popover.find('.am-popover-inner').empty().
-    html(this.options.content);
+    html(content);
 };
 
 Popover.prototype.events = function() {
@@ -212,17 +221,17 @@ Popover.prototype.events = function() {
 function Plugin(option) {
   return this.each(function() {
     var $this = $(this);
-    var data = $this.data('am.popover');
+    var data = $this.data('amui.popover');
     var options = $.extend({},
       UI.utils.parseOptions($this.attr('data-am-popover')),
       typeof option == 'object' && option);
 
     if (!data) {
-      $this.data('am.popover', (data = new Popover(this, options)));
+      $this.data('amui.popover', (data = new Popover(this, options)));
     }
 
     if (typeof option == 'string') {
-      data[option]();
+      data[option] && data[option]();
     }
   });
 }
@@ -230,8 +239,8 @@ function Plugin(option) {
 $.fn.popover = Plugin;
 
 // Init code
-$(function() {
-  $('[data-am-popover]').popover();
+UI.ready(function(context) {
+  $('[data-am-popover]', context).popover();
 });
 
 $.AMUI.popover = Popover;